Archive timeline Phase 1: year-grouped mobile layout
- Extract /archive's accordion list into src/app/archive/_components/ArchiveTimelineMobile.tsx
- Group experiences by start year (parseStartYear helper sorts newest-first, preserves order within a year)
- Each group: accent year heading + hairline rule extending across the column
- Each card gets a bullet dot prefix on the title row — accent solid for past roles, accent + animate-ping for the current Umbra role
- archive/page.tsx becomes a server component shell — back link + SectionTitle + intro + the timeline. experienceCount drives the SectionTitle meta dynamically
- All existing accordion behaviour preserved (expand/collapse, highlights stagger, skills chips, hover bar)
- Renders at every viewport for now; Phase 2 will add a cinematic desktop variant with a rail and gate this to < md
